Add new installByRevisionOne which implements RefreshConfig.
InstallByRevision is too different from the other configs using
executeOne, thus need a new one: installByRevisionOne.  No base nor
channel are necessary.  Configs using executeOne do not require the
Fields.

Remove base from InstallByRevision args, it's not sent via the api.

Move RefreshConfig and all structs implementing it to a new file for
readability.
