+++ core/charm/repository/charmhub_test.go
@@ -547,7 +547,6 @@ func (refreshConfigSuite) TestRefreshByChannel(c *gc.C) {
 			},
 		}},
 		Context: []transport.RefreshRequestContext{},
-		Fields:  []string{"bases", "download", "id", "revision", "version", "resources"},
 	})
 }
 
@@ -581,7 +580,6 @@ func (refreshConfigSuite) TestRefreshByChannelVersion(c *gc.C) {
 			},
 		}},
 		Context: []transport.RefreshRequestContext{},
-		Fields:  []string{"bases", "download", "id", "revision", "version", "resources"},
 	})
 }
 
@@ -607,7 +605,6 @@ func (refreshConfigSuite) TestRefreshByRevision(c *gc.C) {
 			InstanceKey: instanceKey,
 			Name:        &curl.Name,
 			Revision:    &revision,
-			Base:        &transport.Base{Architecture: "amd64", Name: "ubuntu", Channel: "20.04"},
 		}},
 		Context: []transport.RefreshRequestContext{},
 		Fields:  []string{"bases", "download", "id", "revision", "version", "resources"},
